move rebuilt specific simulation methods to its own util class (sim)
